Flora and fauna put into all tiles
fauna is not positioned inside each tile (but it is in the arrays, the coordinates just haven't been adjusted) because my boss made a comment that I wasn't doing work so I gotta get back to that lmao 

I'm still a bit unsure how to do the visibility thing and I'm just realizing now that I am missing a couple for loops because each flora and fauna array (i.e. mountainFlora, mountainFauna, desertFlora etc.) needs its own for loop to iterate through to make it visible. The logic is that inside the for loop. the object is set to visible. The upper bound of the for loop will be incremented up by one when you press the "flora" button in the edit screen. This will "add" another tree but it is really just making another tree visible by making the upper bound of the for loop larger and thus making a new "random" object visible. The issue is that the objects are hard coded into the different environments so there might be a situation in which you click the add flora button and nothing shows up because the flora you're adding is not visible in your world. This could be solved with a line of logic that says "if (container that object is located in) is not shown, increment i +1" this would make it so that it would keep incrementing until it shows an object properly. This would make it so that when you eventually unlock another tile, it would already be populated by flora and fauna from those other tiles. facetime/call me if this makes no sense!!!!! I will be on campus tonight and can probably come help for a bit if need be!